[FIX] Correções de problemas pós atualização.
