+++ src/Infra/Repositories/MainTaskRepository.cs
@@ -8,6 +8,7 @@ public interface IMainTaskRepository
 {
     MainTask Create(MainTask mainTask);
     List<MainTask> Get(int userId);
+    MainTask? Find(int mainTaskId);
     MainTask Update(MainTask mainTask, int mainTaskId);
     void Delete(int mainTaskId);
 }
@@ -31,6 +32,11 @@ public class MainTaskRepository : IMainTaskRepository
     public List<MainTask> Get(int userId)
     {
         return _myDBContext.MainTasks.Where(x => x.UserId == userId).ToList();
+    }
+
+    public MainTask? Find(int mainTaskId)
+    {
+        return _myDBContext.MainTasks.Find(mainTaskId);
     }
 
     public MainTask Update(MainTask mainTaskUpdate, int mainTaskId)
